refactor: Move chartConfig fetch & save logic into a new folder
